scripts/btc-positions/requirements.txt
attrs==20.3.0
boto3==1.17.51
botocore==1.20.51
boxsdk==2.11.0
certifi==2020.12.5
chardet==4.0.0
elasticsearch==7.12.0
idna==2.10
jmespath==0.10.0
ndjson==0.3.1
numpy==1.20.2
pandas==1.2.3
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.9.1
s3transfer==0.3.7
six==1.15.0
urllib3==1.26.4
wrapt==1.12.1
